Move Store classes to own package
